% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SphereOptimize.R
\name{from.Sphere}
\alias{from.Sphere}
\title{Converting spherical coordinates to Cartesian coordinates}
\usage{
from.Sphere(theta)
}
\arguments{
\item{theta}{A list of angles. The first item should be between 0 to pi, and the following items
should be between 0 to 2*pi.}
}
\value{
A vector of the corresponding Cartesian coordinates.
}
\description{
The function from.Sphere convert a list of angles representing a point
on a unit sphere to the corresponding Cartesian coordinates.
}
\examples{
from.Sphere(c(pi/3, pi/4, pi/5))
}
